+++ config/rig_tests/test.exs
@@ -10,7 +10,7 @@ config :rig, RigTests.Proxy.ResponseFrom.KafkaTest,
   # The list of brokers, given by a comma-separated list of host:port items:
   brokers: {:system, :list, "KAFKA_BROKERS", []},
   serializer: {:system, "KAFKA_SERIALIZER", nil},
-  schema_registry_host: {:system, "KAFKA_SCHEMA_REGISTRY_HOST", "localhost:8081"},
+  schema_registry_host: {:system, "KAFKA_SCHEMA_REGISTRY_HOST", nil},
   # The list of topics to consume messages from:
   consumer_topics: {:system, :list, "KAFKA_SOURCE_TOPICS", []},
   # If KAFKA_SSL_ENABLED=0, the KAFKA_SSL_* settings are ignored; otherwise, they're required.
@@ -30,7 +30,7 @@ config :rig, RigTests.Proxy.PublishToEventStream.KafkaTest,
   # The list of brokers, given by a comma-separated list of host:port items:
   brokers: {:system, :list, "KAFKA_BROKERS", []},
   serializer: {:system, "KAFKA_SERIALIZER", nil},
-  schema_registry_host: {:system, "KAFKA_SCHEMA_REGISTRY_HOST", "localhost:8081"},
+  schema_registry_host: {:system, "KAFKA_SCHEMA_REGISTRY_HOST", nil},
   # The list of topics to consume messages from:
   consumer_topics: {:system, :list, "KAFKA_SOURCE_TOPICS", []},
   # If KAFKA_SSL_ENABLED=0, the KAFKA_SSL_* settings are ignored; otherwise, they're required.
@@ -49,7 +49,7 @@ config :rig, RigTests.Avro.AvroTest,
   # The list of brokers, given by a comma-separated list of host:port items:
   brokers: {:system, :list, "KAFKA_BROKERS", []},
   serializer: {:system, "KAFKA_SERIALIZER", nil},
-  schema_registry_host: {:system, "KAFKA_SCHEMA_REGISTRY_HOST", "localhost:8081"},
+  schema_registry_host: {:system, "KAFKA_SCHEMA_REGISTRY_HOST", nil},
   # The list of topics to consume messages from:
   consumer_topics: {:system, :list, "KAFKA_SOURCE_TOPICS", []},
   # If KAFKA_SSL_ENABLED=0, the KAFKA_SSL_* settings are ignored; otherwise, they're required.
@@ -68,7 +68,7 @@ config :rig, RigTests.Proxy.RequestLogger.KafkaTest,
   # The list of brokers, given by a comma-separated list of host:port items:
   brokers: {:system, :list, "KAFKA_BROKERS", []},
   serializer: {:system, "KAFKA_SERIALIZER", nil},
-  schema_registry_host: {:system, "KAFKA_SCHEMA_REGISTRY_HOST", "localhost:8081"},
+  schema_registry_host: {:system, "KAFKA_SCHEMA_REGISTRY_HOST", nil},
   # The list of topics to consume messages from:
   consumer_topics: ["rig-request-log"],
   # If KAFKA_SSL_ENABLED=0, the KAFKA_SSL_* settings are ignored; otherwise, they're required.